Crate sc_consensus_epochs

Generic utilities for epoch-based consensus engines.

Modules

migration
Migration types for epoch changes.

Structs

EpochChanges
Tree of all epoch changes across all seen forks. Data stored in tree is the hash and block number of the block signaling the epoch change, and the epoch that was signalled at that block.

EpochHeader
Header of epoch data, consisting of start and end slot.

EpochIdentifier
Epoch identifier.

HeaderBackendDescendentBuilder
Wrapper to get around unconstrained type errors when implementing IsDescendentOfBuilder for header backends.

IncrementedEpoch
A fresh, incremented epoch to import into the underlying fork-tree.

Enums

EpochIdentifierPosition
Position of the epoch identifier.

PersistedEpoch
Persisted epoch stored in EpochChanges.

PersistedEpochHeader
Persisted epoch header stored in ForkTree.

ViableEpoch
The viable epoch under which a block can be verified.

ViableEpochDescriptor
Descriptor for a viable epoch.

Traits

Epoch
Epoch data, distinguish whether it is genesis or not.

IsDescendentOfBuilder
A builder for is_descendent_of functions.

Functions

descendent_query
Produce a descendent query object given the client.

Type Definitions

EpochChangesFor
Type alias to produce the epoch-changes tree from a block type.

SharedEpochChanges
A shared epoch changes tree.