Trait sc_consensus_epochs::IsDescendentOfBuilder

pub trait IsDescendentOfBuilder<Hash> {
    type Error: Error;
    type IsDescendentOf: Fn(&Hash, &Hash) -> Result<bool, Self::Error>;
    fn build_is_descendent_of(
        &self, 
        current: Option<(Hash, Hash)>
    ) -> Self::IsDescendentOf;
}

A builder for is_descendent_of functions.

Associated Types

type Error: Error

The error returned by the function.

type IsDescendentOf: Fn(&Hash, &Hash) -> Result<bool, Self::Error>

A function that can tell you if the second parameter is a descendent of the first.

Required methods

fn build_is_descendent_of(
&self,
current: Option<(Hash, Hash)>
) -> Self::IsDescendentOf

Build an is_descendent_of function.

The current parameter can be Some with the details a fresh block whose details aren’t yet stored, but its parent is.

The format of current when Some is (current, current_parent).